Move _ModelContext
